Did some refactoring.
