+++ src/gantry/_types.py
@@ -1,9 +1,4 @@
 from pathlib import Path
-from typing import NamedTuple
 
 
 PathLike = Path | str
-
-
-class ProgramOptions(NamedTuple):
-    services_path: Path
