+++ src/gantry/cli/_main.py
@@ -1,10 +1,10 @@
 import click
 
 from . import build, configure, schemas
-from ._util import configure_logger
+from ._common import ProgramOptions, configure_logger
 
 from .. import __version__
-from .._types import Path, ProgramOptions
+from .._types import Path
 
 
 @click.group()
