+++ src/gantry/cli/build.py
@@ -2,8 +2,8 @@ import datetime
 
 import click
 
-from ._util import load_service_group
-from .._types import Path, ProgramOptions
+from ._common import ProgramOptions, load_service_group
+from .._types import Path
 from ..targets import ImageTarget
