+++ src/gantry/cli/configure.py
@@ -1,6 +1,6 @@
 import click
 
-from ._util import load_service_group
+from ._common import load_service_group
 
 from .._types import Path
 from ..exceptions import ComposeServiceBuildError
